fix: created separate test files
